Budget null 문제

+++ src/main/java/com/hana/ddok/budget/dto/BudgetByCategoryUpdateReq.java
@@ -1,15 +1,15 @@
 package com.hana.ddok.budget.dto;
 
 public record BudgetByCategoryUpdateReq(
-        Integer shopping,
-        Integer food,
-        Integer traffic,
-        Integer hospital,
-        Integer fee,
-        Integer education,
-        Integer leisure,
-        Integer society,
-        Integer daily,
-        Integer overseas
+        Long shopping,
+        Long food,
+        Long traffic,
+        Long hospital,
+        Long fee,
+        Long education,
+        Long leisure,
+        Long society,
+        Long daily,
+        Long overseas
 ) {
 }
